+++ resolwe/flow/tests/test_executors.py
@@ -4,13 +4,15 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import mock
 import os
 import shlex
+import six
 import subprocess
 import unittest
 
 from django.conf import settings
 from django.test import override_settings
 
-from ..executors.docker import FlowExecutor
+from resolwe.flow.executors.docker import FlowExecutor
+from resolwe.flow.executors import BaseFlowExecutor
 
 try:
     import builtins  # py3
@@ -72,3 +74,34 @@ class DockerExecutorTestCase(unittest.TestCase):
                 executor.run('no_data_id', script, verbosity=0)
 
             self.assertEqual(count['running'], 1)
+
+
+class GetToolsTestCase(unittest.TestCase):
+    @mock.patch('resolwe.flow.executors.apps')
+    @mock.patch('resolwe.flow.executors.os')
+    @mock.patch('resolwe.flow.executors.settings')
+    def test_get_tools(self, settings_mock, os_mock, apps_mock):
+        apps_mock.get_app_configs.return_value = [
+            mock.MagicMock(path='/resolwe/test_app1'),
+            mock.MagicMock(path='/resolwe/test_app2'),
+        ]
+        os_mock.path.join = os.path.join
+        os_mock.path.isdir.side_effect = [False, True]
+        settings_mock.RESOLWE_CUSTOM_TOOLS_PATHS = ['/custom_tools']
+
+        base_executor = BaseFlowExecutor()
+        tools_list = base_executor.get_tools()
+
+        self.assertEqual(len(tools_list), 2)
+        self.assertIn('/resolwe/test_app2/tools', tools_list)
+        self.assertIn('/custom_tools', tools_list)
+
+    @mock.patch('resolwe.flow.executors.apps')
+    @mock.patch('resolwe.flow.executors.settings')
+    def test_not_list(self, settings_mock, apps_mock):
+        apps_mock.get_app_configs.return_value = []
+        settings_mock.RESOLWE_CUSTOM_TOOLS_PATHS = '/custom_tools'
+
+        base_executor = BaseFlowExecutor()
+        with six.assertRaisesRegex(self, KeyError, 'setting must be a list'):
+            base_executor.get_tools()